# sdc-api/src/main/java/br/com/conductor/sbc/util/Response.java
package br.com.conductor.sbc.util;

import javax.persistence.Transient;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Response {

	@Transient
	private String msg;

}