# sdc-api/src/main/resources/json/swagger.json
{"swagger":"2.0","info":{"description":"API desenvolvida para a 15Âª semanda da computaÃ§Ã£o da UFPB em 2016","version":"1.0.0","title":"API SDC 2016","termsOfService":"http://pierlabs.io/terms/","contact":{"name":"PIER Labs","url":"http://pierlabs.io","email":"pierlabs@conductor.com.br"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost","basePath":"/","tags":[{"name":"Cartao","description":"OperaÃ§Ãµes com CartÃµes"},{"name":"Conta","description":"OperaÃ§oes com Contas"}],"paths":{"/v1/contas":{"get":{"tags":["Conta"],"summary":"Retorna todas as contas","description":"Retorna todas as contas","operationId":"getAllUsingGET_1","consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Conta"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]},"post":{"tags":["Conta"],"summary":"Cria uma conta","description":"Cria uma conta","operationId":"createUsingPOST_1","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"conta","description":"conta","required":true,"schema":{"$ref":"#/definitions/Conta"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Conta"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]},"put":{"tags":["Conta"],"summary":"Atualiza uma conta","description":"Atualiza uma conta","operationId":"updateUsingPUT_1","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"conta","description":"conta","required":true,"schema":{"$ref":"#/definitions/Conta"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Conta"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes":{"get":{"tags":["Cartao"],"summary":"Retona todos os cartÃ£o de uma conta","description":"Consulte todos os carteos de uma determinada conta.","operationId":"getAllUsingGET","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Cartao"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]},"post":{"tags":["Cartao"],"summary":"Cria um cartÃ£o","description":"VocÃª pode criar um cartÃ£o ligado hÃ¡ uma determinada conta.","operationId":"createUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"in":"body","name":"cartao","description":"CartÃ£o a ser criado","required":true,"schema":{"$ref":"#/definitions/Cartao"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cartao"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]},"put":{"tags":["Cartao"],"summary":"Atualiza um cartÃ£o","description":"Atualiza um determinado cartÃ£o","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"in":"body","name":"cartao","description":"CartÃ£o a ser atualizado","required":true,"schema":{"$ref":"#/definitions/Cartao"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cartao"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}":{"get":{"tags":["Cartao"],"summary":"Retorna um cartÃ£o","description":"Consulte um determinado cartÃ£o","operationId":"getOneUsingGET","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"ID do CartÃ£o","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cartao"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/bloquear":{"put":{"tags":["Cartao"],"summary":"Bloquear um cartÃ£o.","description":"Bloquear um cartÃ£o.","operationId":"bloquearUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"CartÃ£o a ser bloqueado","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/cancelar":{"delete":{"tags":["Cartao"],"summary":"Cancelar um cartÃ£o","description":"Cancelar um determinado cartÃ£o","operationId":"cancelarUsingDELETE","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"CartÃ£o a ser cancelado","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/creditar":{"put":{"tags":["Cartao"],"summary":"Creditar dinheiro em um cartÃ£o","description":"Com esse mÃ©todo Ã© possivel creditar qualquer valor em um cartÃ£o.","operationId":"creditarUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"ID do Cartao a ser creditado","required":true,"type":"integer","format":"int64"},{"in":"body","name":"valor","description":"Valor a ser creditado","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/desbloquear":{"put":{"tags":["Cartao"],"summary":"Desbloquear um cartÃ£o.","description":"Desbloquear um determinado cartÃ£o.","operationId":"desbloquearUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"CartÃ£o a ser desbloqueado","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/extratos":{"post":{"tags":["Cartao"],"summary":"Retonar os extratos de transaÃ§oes do cartÃ£o","description":"Retorna os extratos de todas as transaÃ§oes de um determinado cartÃ£o","operationId":"extratosUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"ID do cartÃ£o","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Transacao"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/limite":{"get":{"tags":["Cartao"],"summary":"Retorna o limite do cartÃ£o","description":"Consultar o limite de um determinado cartÃ£o","operationId":"limiteUsingGET","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"ID do Cartao a ser creditado","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Limite"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/transacionar":{"put":{"tags":["Cartao"],"summary":"Transacionar valores","description":"Transacionar algum valor utilizando um determinado cartÃ£o","operationId":"transacionarUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"ID do Cartao a ser creditado","required":true,"type":"integer","format":"int64"},{"in":"body","name":"valor","description":"Valor da transaÃ§Ã£o","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{idConta}/cartoes/{idCartao}/transferir":{"post":{"tags":["Cartao"],"summary":"Transferir valores entre dois cartÃµes distintos","description":"Com esse mÃ©todo Ã© permitido transferir qualquer valores entre dois cartÃµes. Basta o cartÃ£o possuir limite e estar desbloqueado.","operationId":"transferirUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"idConta","in":"path","description":"ID da Conta do cartÃ£o de origem","required":true,"type":"integer","format":"int64"},{"name":"idCartao","in":"path","description":"ID do cartÃ£o de origem","required":true,"type":"integer","format":"int64"},{"name":"idCartaoDestino","in":"query","description":"ID do cartÃ£o de destino","required":true,"type":"integer","format":"int64"},{"in":"body","name":"valor","description":"Valor a ser transferido","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]}},"/v1/contas/{id}":{"get":{"tags":["Conta"],"summary":"Retorna uma conta","description":"Retorna uma conta","operationId":"getOneUsingGET_1","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Conta"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"access_token":["global"]}]},"delete":{"tags":["Conta"],"summary":"Deleta uma conta","description":"Deleta um cartÃ£o","operationId":"deleteUsingDELETE","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}},"security":[{"access_token":["global"]}]}}},"securityDefinitions":{"access_token":{"type":"apiKey","name":"access_token","in":"header"},"client_id":{"type":"apiKey","name":"client_id","in":"header"}},"definitions":{"Response":{"type":"object"},"Limite":{"type":"object","properties":{"valor":{"type":"number","format":"double"}}},"Transacao":{"type":"object","properties":{"dataTransacao":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"valor":{"type":"number","format":"double"}}},"Conta":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"}}},"Cartao":{"type":"object","properties":{"cvv":{"type":"string"},"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"numero":{"type":"string"},"senha":{"type":"string"},"status":{"type":"string","enum":["ATIVO","CANCELADO","BLOQUEADO"]}}}}}